all: move InputParser to nwg_classes.
