+++ bar/bar.cc
@@ -10,6 +10,7 @@
 #include <sys/time.h>
 
 #include "nwg_tools.h"
+#include "nwg_classes.h"
 #include "on_event.h"
 #include "bar.h"
