+++ common/meson.build
@@ -1,6 +1,7 @@
 sources = files(
 	'nwg_tools.cc',
-	'on_event.cc'
+	'on_event.cc',
+	'nwg_classes.cc'
 )
 
 nwg_inc = include_directories('.')
